Apply ACL's on s3cmd get
s3cmd wasn't applying ACL's on the file copied from s3. Hence the
'permissions' information wouldn't match the one which is on s3. For
e.g. This affects the user when he stores an 'executable' and downloads
it and cannot run it.

This patch fixes the issue by applying the ACL attributes from s3 after
the file is downloaded using s3cmd get.